ref: change project architecture
