+++ lib/models/search_result_model.dart
