+++ lib/repositories/user_repository_impl.dart
@@ -1,14 +1,8 @@
 import 'dart:convert';
 
 import 'package:shared_preferences/shared_preferences.dart';
-import 'package:violin/domain/user/user_model.dart';
-
-abstract class UserRepository {
-  Future<void> saveUser(UserModel user);
-  Future<UserModel?> getUser();
-  Future<void> updateUser(UserModel updatedUser);
-  Future<void> deleteUser();
-}
+import 'package:violin/models/user_model.dart';
+import 'package:violin/repositories/user_repository.dart';
 
 class UserRepositoryImpl implements UserRepository {
   static const String _userKey = 'current_user';
